Code structure reorganization
