+++ src/components/Account Route/Account.js
@@ -1,11 +1,10 @@
-import Layout from "./Layout";
-import {UserAuth} from "../context/AuthContext";
+import Layout from "../Layout";
 import {MdPerson} from "react-icons/md";
 import {IoCloseCircleOutline} from "react-icons/io5";
 import React, {useEffect, useState} from "react"
 import {Box, Tab, Tabs, Typography} from "@mui/material";
 import {collection, deleteDoc, doc, getDocs} from "firebase/firestore";
-import {auth, db} from "../firebase";
+import {auth, db} from "../../firebase";
 import {useHistory} from "react-router-dom";
 
 interface TabPanelProps {
